Leaf Collectible
The leaf game object was added to the level. With the use of HealthCollectible Script, when the player collides with the leaf it will be collected and the health count will go up by 1. With the same logic when the player falls into water or the "damage zone" one leaf will be deducted from overall health.