Audio for Level 1
In this commit a SoundManager was made and a script for it to manage the sound for the scene.
4 mp3 files were imported - background song, jump sound, collectible sound and damage zone sound.
Each of these sounds were added as private AudioClip to the respective script and played through SoundManager ex during collision.
The LoadingManager was added in advance for when the next scene gets loaded the background music doesn't abruptly stop or double play.