Pause Menu
Pause menu with options to resume, change volume, change music volume, restart or quit the game.
UICanvas with UI image for the pause screen and texts for the options + UI image for the selection arrow.
UIManager script contains all methods for the options, SelectionArrow script makes sure the player can go from option to an option and select it.